checking in tar.gz in archive before releting directory

+++ Ramses_analysis/gadi_archive_sim_data.py
@@ -17,5 +17,9 @@ for data_dir in data_dirs:
         subprocess.run('tar -cvfr output_'+("%05d" % dir_it)+'.tar.gz data/output_'+("%05d" % dir_it)+'/', shell=True)
         subprocess.run('mdss put output_'+("%05d" % dir_it)+'.tar.gz rlk100/2023_paper/Simulation_data/', shell=True)
         subprocess.run('rm output_'+("%05d" % dir_it)+'.tar.gz', shell=True)
-        subprocess.run('rm -rf data/output_'+("%05d" % dir_it)+'/', shell=True)
+        result = subprocess.run('mdss ls rlk100/2023_paper/Simulation_data/', shell=True, capture_output=True, text=True)
+        output = result.stdout.strip()
+        output = output.split('\n')
+        if 'output_'+("%05d" % dir_it)+'.tar.gz' in output:
+            subprocess.run('rm -rf data/output_'+("%05d" % dir_it)+'/', shell=True)
